+++ midiscripter/base/__init__.py
@@ -1,3 +1,3 @@
 from .msg_base import Msg
-from .port_base import MultiPort, CallOn, ports_opened
+from .port_base import IoPort, CallOn, ports_opened
 from .match_conditions import Not, Glob, Regex